# docs/core/openapi.yaml
openapi: 3.1.0
info:
  title: ecommerce-aigc-video API
  version: 2.0.0
  description: >
    V2 目标后端契约。全部路由前缀 /api，URL 无版本号，全站无鉴权。
    工作区级 agent 模块采用 propose -> approve 语义：propose 只生成待审 artifact，
    approve 才变为当前生效 artifact；下游只读取 approved/current。
    approve 不级联重置下游，shotprompt approve 也不重建 shots。
    新的分镜链路由显式创建 shot set 触发。上游变更通过 upstreamChanged 提示表达，
    不代表 stale，也不删除候选或选择。图像/视频选择写入 image_select_artifacts /
    video_select_artifacts，每个 shot 只有一个当前 selected image/video，重复选择用 UPSERT 覆盖。
    业务语义见 docs/core/interface.md，架构见 docs/core/arc_v2.md，数据模型见 docs/core/erd.md。
servers:
  - url: http://localhost:3000
    description: 本地开发（SERVER_PORT 默认 3000）
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string }
...
    ShotId:
      name: shotId
      in: path
      required: true
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1 }
  responses:
    Error:
      description: 业务错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
...
    ShotStatus:
      type: string
      enum:
        - DRAFT
        - IMAGE_PROMPT_PROPOSING
        - IMAGE_PROMPT_READY
        - IMAGE_GENERATING
        - IMAGE_CANDIDATES_READY
        - IMAGE_SELECTED
        - VIDEO_SCRIPT_PROPOSING
        - VIDEO_SCRIPT_READY
        - VIDEO_GENERATING
        - VIDEO_CANDIDATES_READY
        - VIDEO_SELECTED
        - FAILED
    UpstreamChange:
      type: object
      required: [upstreamChanged, changedSources]
      properties:
        upstreamChanged: { type: boolean }
        changedSources:
          type: array
          items: { type: string }
        message: { type: string }
...
    ShotRequirements:
      type: object
      required: [shotImage, shotVideo]
      properties:
        shotImage:
          $ref: "#/components/schemas/ShotImageRequirement"
        shotVideo:
          $ref: "#/components/schemas/ShotVideoRequirement"
        sourceShotPromptArtifactId: { type: string }
    Shot:
      type: object
      required: [id, workspaceId, shotSetId, orderIndex, title, status]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        shotSetId: { type: string }
        orderIndex:
          type: integer
          description: >-
            0-based workflow order derived from shot array position; not the
            provider-facing shots[].index.
        title: { type: string }
        objective: { type: string }
        defaultDurationSec: { type: integer }
        status:
          $ref: "#/components/schemas/ShotStatus"
        requirements:
          $ref: "#/components/schemas/ShotRequirements"
        selectedImage:
          anyOf:
            - $ref: "#/components/schemas/ImageSelectArtifact"
            - type: "null"
        selectedVideo:
          anyOf:
            - $ref: "#/components/schemas/VideoSelectArtifact"
            - type: "null"
        upstream:
          $ref: "#/components/schemas/UpstreamChange"
        nextAction:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ShotAssetRef:
      type: object
      required: [id, shotId, assetId, role, weight, position]
      properties:
        id: { type: string }
        shotId: { type: string }
        assetId: { type: string }
        role:
          type: string
          enum:
            [
              product_identity,
              reference_style,
              reference_scene,
              first_frame_hint,
              other
            ]
        weight: { type: number, default: 1 }
        position: { type: integer, minimum: 0 }
        ref: { type: string }
        url: { type: string }
        metadata:
          type: object
          additionalProperties: true
    PatchShotAssetRefsRequest:
      type: object
      required: [refs]
      properties:
        refs:
          type: array
          items:
            type: object
            required: [assetId, role]
            properties:
              assetId: { type: string }
              role:
                type: string
                enum:
                  [
                    product_identity,
                    reference_style,
                    reference_scene,
                    first_frame_hint,
                    other
                  ]
              weight: { type: number, minimum: 0, maximum: 1 }
...
    ImagePromptArtifact:
      type: object
      properties:
        id: { type: string }
        shotId: { type: string }
        version: { type: integer }
        status: { type: string, enum: [DRAFT, ACTIVE, APPROVED, STALE, ARCHIVED] }
        promptText: { type: string }
        negativePrompt: { type: string }
        referenceAssetIds:
          type: array
          items: { type: string }
        promptJson:
          $ref: "#/components/schemas/ImagePromptJson"
        sourceFingerprint:
          $ref: "#/components/schemas/SourceFingerprint"
        promptAssembly:
          $ref: "#/components/schemas/PromptAssemblyMetadata"
        createdBy: { type: string, enum: [agent, user, system] }
        agentName: { type: string, nullable: true }
        baseArtifactId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    ImagePromptRegenerateRequest:
      type: object
      required: [baseArtifactId]
      additionalProperties: false
      properties:
        baseArtifactId: { type: string }
        userDirection:
          type: string
          description: 用户对本次重新生成的补充反馈；最终 prompt 仍由后端基于 shotImage 和 shotVideo 重新装配。
...
    VideoScriptArtifact:
      type: object
      properties:
        id: { type: string }
        shotId: { type: string }
        version: { type: integer }
        status: { type: string, enum: [DRAFT, ACTIVE, APPROVED, STALE, ARCHIVED] }
        durationSec: { type: number }
        scriptJson:
          $ref: "#/components/schemas/VideoScriptJson"
        providerPrompt: { type: string }
        basedOnImageCandidateId: { type: string, nullable: true }
        basedOnPrevImageCandidateId: { type: string, nullable: true }
        basedOnNextImageCandidateId: { type: string, nullable: true }
        sourceFingerprint:
          $ref: "#/components/schemas/SourceFingerprint"
        promptAssembly:
          $ref: "#/components/schemas/PromptAssemblyMetadata"
        createdAt: { type: string, format: date-time }
    GenerationBatch:
      type: object
      properties:
        id: { type: string }
        workspaceId: { type: string }
        shotId: { type: string }
        status:
          $ref: "#/components/schemas/BatchStatus"
        requestedCount: { type: integer }
        succeededCount: { type: integer }
        failedCount: { type: integer }
        provider: { type: string }
        aspectRatio:
          $ref: "#/components/schemas/AspectRatio"
        providerRequest: { type: object, additionalProperties: true }
        errorMessage: { type: string, nullable: true }
    ImageCandidate:
      type: object
      properties:
        id: { type: string }
        batchId: { type: string }
        workspaceId: { type: string }
        shotId: { type: string }
        imageUrl: { type: string, nullable: true }
        objectKey: { type: string, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        provider: { type: string }
        providerResponse: { type: object, additionalProperties: true }
        status:
          $ref: "#/components/schemas/CandidateStatus"
        errorMessage: { type: string, nullable: true }
    VideoCandidate:
      type: object
      properties:
        id: { type: string }
        batchId: { type: string }
        workspaceId: { type: string }
        shotId: { type: string }
        videoUrl: { type: string, nullable: true }
        objectKey: { type: string, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        durationSec: { type: integer, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        provider: { type: string }
        providerResponse: { type: object, additionalProperties: true }
        status:
          $ref: "#/components/schemas/CandidateStatus"
        errorMessage: { type: string, nullable: true }
...
    ImageSelectionView:
      type: object
      required:
        [shotId, selectedCandidateId, selectedImageUrl, nextShotId, allShotsImageSelected]
      description: >-
        当前 shot 的已确认分镜图视图。重新生成候选图不会清空或替换该选择；
        当最新 round 的 candidates 不包含 selectedCandidateId 时，selectedImageUrl
        仍指向旧轮次已确认图片，供前端继续展示当前图。
      properties:
        shotId: { type: string }
        selectedCandidateId: { type: string }
        selectedImageUrl: { type: string }
        nextShotId: { type: string, nullable: true }
        allShotsImageSelected: { type: boolean }
    VideoSelectionView:
      type: object
      required:
        [shotId, selectedCandidateId, selectedVideoUrl, duration, allShotsVideoSelected]
      description: 当前 shot 的已确认分镜视频视图。
      properties:
        shotId: { type: string }
        selectedCandidateId: { type: string }
        selectedVideoUrl: { type: string }
        duration: { type: number }
        allShotsVideoSelected: { type: boolean }
    ImageRound:
      type: object
      properties:
        artifact:
          $ref: "#/components/schemas/ImagePromptArtifact"
        batch:
          anyOf:
            - $ref: "#/components/schemas/GenerationBatch"
            - type: "null"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/ImageCandidate"
        selection:
          anyOf:
            - $ref: "#/components/schemas/ImageSelectionView"
            - type: "null"
          description: >-
            当前 shot 的已确认分镜图；可能来自旧候选轮次，不要求属于本 round.candidates。
        upstream:
          $ref: "#/components/schemas/UpstreamChange"
        context: {}
    VideoRound:
      type: object
      properties:
        artifact:
          $ref: "#/components/schemas/VideoScriptArtifact"
        batch:
          anyOf:
            - $ref: "#/components/schemas/GenerationBatch"
            - type: "null"
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/VideoCandidate"
        selection:
          anyOf:
            - $ref: "#/components/schemas/VideoSelectionView"
            - type: "null"
        frames:
          type: object
          properties:
            firstFrameUrl: { type: string, nullable: true }
            lastFrameUrl: { type: string, nullable: true }
            firstFrameCandidateId: { type: string, nullable: true }
            lastFrameCandidateId: { type: string, nullable: true }
        upstream:
          $ref: "#/components/schemas/UpstreamChange"
        context: {}
...
paths:
...
  /api/workspaces/{workspaceId}/shots:
    get:
      tags: [Shot]
      summary: 获取 active shot set 下的 shots
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Shots
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Shot"
        "400":
          $ref: "#/components/responses/Error"
  /api/shots/{shotId}:
    get:
      tags: [Shot]
      summary: 获取单个 shot
      parameters:
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Shot
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Shot"
  /api/shots/{shotId}/asset-refs:
    get:
      tags: [Shot]
      summary: 列出 shot 的可编辑素材引用
      parameters:
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Shot asset refs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShotAssetRef"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Shot]
      summary: 替换 shot 的可编辑素材引用
      description: >-
        用请求体顺序替换当前 active shot 的素材引用，position 由数组顺序决定。
        该接口不重跑 image/video 链路，只更新下一次 prompt assembly 读取的引用。
      parameters:
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchShotAssetRefsRequest"
      responses:
        "200":
          description: Updated shot asset refs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShotAssetRef"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/shots/{shotId}/image-prompts/propose:
    post:
      tags: [Shot]
      summary: 运行图像 prompt agent 并创建图像候选 batch
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                userDirection: { type: string }
      responses:
        "200":
          description: Image prompt artifact and batch
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImagePromptArtifact"
                  artifact:
                    $ref: "#/components/schemas/ImagePromptArtifact"
                  batch:
                    $ref: "#/components/schemas/GenerationBatch"
                  candidates:
                    type: array
                    items:
                      $ref: "#/components/schemas/ImageCandidate"
                  usage: { nullable: true }
                  shotStatus:
                    $ref: "#/components/schemas/ShotStatus"
                  nextAction:
                    type: object
                    additionalProperties: true
                  traceId: { type: string }
                  context:
                    type: object
                    additionalProperties: true
  /api/workspaces/{workspaceId}/shots/{shotId}/image-prompts/regenerate:
    post:
      tags: [Shot]
      summary: 用户反馈后重新装配 image prompt 并生成候选图
      description: >-
        基于 base image prompt artifact、当前 shotImage/shotVideo 和用户反馈重新调用
        image-prompt agent，创建新的 user-created image prompt artifact 和 image batch。
        不清空当前 selected image，旧候选仍保留可回看。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImagePromptRegenerateRequest"
      responses:
        "200":
          description: User-feedback image prompt artifact and new batch
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImagePromptArtifact"
                  artifact:
                    $ref: "#/components/schemas/ImagePromptArtifact"
                  batch:
                    $ref: "#/components/schemas/GenerationBatch"
                  candidates:
                    type: array
                    items:
                      $ref: "#/components/schemas/ImageCandidate"
                  shotStatus:
                    $ref: "#/components/schemas/ShotStatus"
                  nextAction:
                    type: object
                    additionalProperties: true
                  traceId: { type: string }
                  context:
                    type: object
                    additionalProperties: true
        "400":
          $ref: "#/components/responses/Error"
  /api/shots/{shotId}/image-prompts:
    get:
      tags: [Shot]
      summary: 列出 shot 图像 prompt artifacts
      parameters:
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Image prompt artifacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ImagePromptArtifact"
  /api/workspaces/{workspaceId}/shots/{shotId}/image-rounds:
    get:
      tags: [Shot]
      summary: 获取图像生成轮次
      description: >-
        按 image prompt artifact 聚合候选图轮次。selection 表示当前 shot 已确认分镜图；
        重新生成候选图不会清空该选择，所以最新 round 的 candidates 仍在生成时，
        客户端应继续用 selection.selectedImageUrl 展示当前图。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Image rounds
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ImageRound"
  /api/workspaces/{workspaceId}/shots/{shotId}/image-candidates/select:
    post:
      tags: [Shot]
      summary: 选择当前图像候选
      description: UPSERT image_select_artifacts，不 stale 未选候选。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - required: [candidateId]
                - required: [imageCandidateId]
              properties:
                candidateId: { type: string }
                imageCandidateId: { type: string }
                imageGenerationBatchId: { type: string }
      responses:
        "200":
          description: Current image selection
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImageSelectionView"
                  shotId: { type: string }
                  selectedCandidateId: { type: string }
                  selectedImageUrl: { type: string }
                  nextShotId: { type: string, nullable: true }
                  allShotsImageSelected: { type: boolean }
                  shotStatus:
                    $ref: "#/components/schemas/ShotStatus"
                  nextAction:
                    type: object
                    additionalProperties: true
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/shots/{shotId}/video-scripts/propose:
    post:
      tags: [Shot]
      summary: 运行视频脚本 agent，异步入队每候选一个视频生成 job
      description: >
        写 ACTIVE video_script_artifacts 后，每个候选入队一个 generate_video_candidate
        job 并立即返回 PENDING batch + PENDING candidates；shot 进入 VIDEO_GENERATING。
        客户端轮询 video-rounds 直到 VIDEO_CANDIDATES_READY。video 同时在飞的 Seedance
        调用数被进程级信号量限制为 VIDEO_PROVIDER_CONCURRENCY，并对 429/限流做退避重试。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                userDirection: { type: string }
      responses:
        "200":
          description: Video script artifact plus a PENDING batch with PENDING candidates (poll video-rounds)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/VideoScriptArtifact"
                  artifact:
                    $ref: "#/components/schemas/VideoScriptArtifact"
                  batch:
                    $ref: "#/components/schemas/GenerationBatch"
                  candidates:
                    type: array
                    items:
                      $ref: "#/components/schemas/VideoCandidate"
                  shotStatus:
                    $ref: "#/components/schemas/ShotStatus"
                  nextAction:
                    type: object
                    additionalProperties: true
                  traceId: { type: string }
                  context:
                    type: object
                    additionalProperties: true
                  frames:
                    type: object
                    properties:
                      firstFrameCandidateId: { type: string }
                      lastFrameCandidateId: { type: string, nullable: true }
                      firstFrameUrl: { type: string, nullable: true }
                      lastFrameUrl: { type: string, nullable: true }
                  poll:
                    type: object
                    properties:
                      videoRoundsUrl: { type: string }
                      batchId: { type: string }
        "400":
          $ref: "#/components/responses/Error"
  /api/shots/{shotId}/video-scripts:
    get:
      tags: [Shot]
      summary: 列出 shot 视频脚本 artifacts
      parameters:
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Video script artifacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/VideoScriptArtifact"
  /api/workspaces/{workspaceId}/shots/{shotId}/video-rounds:
    get:
      tags: [Shot]
      summary: 获取视频生成轮次
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      responses:
        "200":
          description: Video rounds
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/VideoRound"
  /api/workspaces/{workspaceId}/shots/{shotId}/video-candidates/select:
    post:
      tags: [Shot]
      summary: 选择当前视频候选
      description: UPSERT video_select_artifacts，不 stale 未选候选。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - required: [candidateId]
                - required: [videoCandidateId]
              properties:
                candidateId: { type: string }
                videoCandidateId: { type: string }
                videoGenerationBatchId: { type: string }
      responses:
        "200":
          description: Current video selection
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/VideoSelectionView"
                  shotId: { type: string }
                  selectedCandidateId: { type: string }
                  selectedVideoUrl: { type: string }
                  duration: { type: number }
                  allShotsVideoSelected: { type: boolean }
                  shotStatus:
                    $ref: "#/components/schemas/ShotStatus"
                  nextAction:
                    type: object
                    additionalProperties: true
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/shots/{shotId}/retry:
    post:
      tags: [Shot]
      summary: 对当前 active prompt/script 重试生成 batch
      parameters:
        - $ref: "#/components/parameters/ShotId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [what]
              properties:
                what: { type: string, enum: [image_batch, video_batch] }
      responses:
        "200":
          description: Generation batch
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/GenerationBatch"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/workspaces/{workspaceId}/shot-workflow-status:
    get:
      tags: [Shot]
      summary: 获取 active shot set 的整体工作流状态
      description: Workspace 存在但尚未生成 active shot set 时返回空 shots 和 canComposeFinalVideo=false；shot 级操作仍要求 active shot set。
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workflow status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      workspaceId: { type: string }
                      canComposeFinalVideo: { type: boolean }
                      shots:
                        type: array
                        items:
                          type: object
                          properties:
                            shotId: { type: string }
                            orderIndex: { type: integer }
                            status: { type: string }
                            nextAction: { type: string }
                            activeImagePromptArtifactId:
                              type: string
                              nullable: true
                            selectedImageId:
                              type: string
                              nullable: true
                            selectedImageUrl:
                              type: string
                              nullable: true
                              description: >-
                                当前 shot 已选分镜图候选的图片 URL；未选择或候选尚无 URL 时为 null。
                                供创作审核台分镜列表直接渲染缩略图，无需逐 shot 调用 image-rounds。
                            activeVideoScriptArtifactId:
                              type: string
                              nullable: true
                            selectedVideoId:
                              type: string
                              nullable: true
                            activeImageBatchId:
                              type: string
                              nullable: true
                            activeVideoBatchId:
                              type: string
                              nullable: true
                            upstream:
                              $ref: "#/components/schemas/UpstreamChange"
...